backend/requirements.txt
fastapi
uvicorn
torch
torchvision
ultralytics
opencv-python
pillow
python-multipart
